+++ app/src/main/java/com/example/sergey/shlypa2/screens/game/GameFragment.kt
@@ -1,4 +1,4 @@
-package com.example.sergey.shlypa2.ui.fragments
+package com.example.sergey.shlypa2.screens.game
 
 
 import android.animation.ValueAnimator
@@ -19,7 +19,6 @@ import com.example.sergey.shlypa2.beans.Word
 import com.example.sergey.shlypa2.utils.Functions
 import com.example.sergey.shlypa2.extensions.hide
 import com.example.sergey.shlypa2.extensions.show
-import com.example.sergey.shlypa2.screens.game.RoundViewModel
 import com.github.florent37.kotlin.pleaseanimate.please
 import kotlinx.android.synthetic.main.fragment_game.*
 import timber.log.Timber
