+++ app/src/main/res/layout/fragment_game.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="com.example.sergey.shlypa2.ui.fragments.GameFragment">
+    tools:context="com.example.sergey.shlypa2.screens.game.GameFragment">
