+++ app/src/main/res/layout/fragment_team_hint.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.sergey.shlypa2.ui.fragments.TeamHintFragment">
+    tools:context="com.example.sergey.shlypa2.screens.game.TeamHintFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvHintTeam_and_Result"
