+++ app/src/main/res/layout/fragment_turn_start.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.sergey.shlypa2.ui.fragments.TurnStartFragment">
+    tools:context="com.example.sergey.shlypa2.screens.game.TurnStartFragment">
 
 
     <TextView
